Refactor: Extract imperative custom cursor logic into dedicated component
